Refactor query transformer and fix failing chained transform
So the big change here is that we no longer require transformers to
invoke continuation.transform() if you're returning a field on
NadelTransformFieldResult.newField

The issue is that NadelTransformFieldResult.newField can be transformed
multiple times, so it shouldn't be until after the last transform that
we actually transform the children. We can't have every transform
transforming the field children.

Note that if you're returning a field in
NadelTransformFieldResult.artificialFields then you MUST invoke
what used to be continuation.transform() IF APPLICABLE

The issue with artificialFields is that sometimes they take real fields
from the overall query, meaning they must be transformed. But sometimes
the fields are completely made up, in which case it does NOT make sense
for the NadelQueryTransformer to do work on it. Because of this, we do
NOT know how to handle fields returned in artificialFields

Thus, the transforms themselves must handle it.